half precision on cpu

+++ aaas/model_utils.py
@@ -16,7 +16,7 @@ def get_model(model_class, model_id):
         cache_dir="./model_cache",
         load_in_8bit=False,
         device_map="auto",
-        torch_dtype=torch.float16 if torch.cuda.is_available() else torch.float32,
+        torch_dtype=torch.float16 if torch.cuda.is_available() else torch.bfloat16,
     )
 
     return model
@@ -37,7 +37,7 @@ def get_peft_model(peft_model_id, model_class) -> peft.PeftModel:
         cache_dir="./model_cache",
         load_in_8bit=False,
         device_map="auto",
-        torch_dtype=torch.float16 if torch.cuda.is_available() else torch.float32,
+        torch_dtype=torch.float16 if torch.cuda.is_available() else torch.bfloat16,
     )
     model = peft.PeftModel.from_pretrained(
         model,
@@ -45,7 +45,7 @@ def get_peft_model(peft_model_id, model_class) -> peft.PeftModel:
         cache_dir="./model_cache",
         load_in_8bit=False,
         device_map="auto",
-        torch_dtype=torch.float16 if torch.cuda.is_available() else torch.float32,
+        torch_dtype=torch.float16 if torch.cuda.is_available() else torch.bfloat16,
     )
     model = model.eval()
